requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
stripe==7.6.0 (0k instalado)
openai==1.3.5 (0k instalado)
httpx==0.25.1 (0k instalado)
pydantic==2.5.0 (0k instalado)
pydantic-settings==2.1.0 (0k instalado)
email-validator==2.1.0 (0k instalado) WARNING: The candidate selected for download or install is a yanked version: 'email-validator' candidate (version 2.1.0 at https://files.pythonhosted.org/packages/90/41/4767ff64e422734487a06384a66e62615b1f5cf9cf3b23295e22d3ecf71
1/email_validator-2.1.0-py3-none-any.whl (from https://pypi.org/simple/email-validator/) (requires-python:>=3.7))

celery==5.3.4 (0k instalado)
redis==5.0.1 (0k instalado)
jinja2==3.1.2 (0k instalado)
aiofiles==23.2.1 (0k instalado)
Pillow==10.1.0 (0k instalado)
python-jwt==4.0.0 (0k instalado)
bcrypt==4.1.2